refactor: use generics
By using a conversion method named after the generic type
(one for each user-defined type of the generic with type parameters),
all subtypes of the generic type have access
to the generic type's methods.

This way we get something like generic methods.
The only annoyance is the extra method call between the generic method
and the user-defined type that converts the user-defined
type to the generic with type parameters.